Resolve common folder missing dependency
