--- common/models/sheet/inventory.ts
@@ -1,6 +1,0 @@
-import { Slottable } from "./itemDetails";
-
-export class Inventory {
-    // scared of thinking about how to do this
-    constructor(public items: Slottable[]) {}; 
-}
