+++ common/src/models/inventory.ts
@@ -1,0 +1,6 @@
+import { Slottable } from "./itemDetails";
+
+export class Inventory {
+    // scared of thinking about how to do this
+    constructor(public items: Slottable[]) {}; 
+}
